fixed channels for 2.4 and 5gh of country Japan(Android)
