+++ tests/e2e/interOp/validation_of_operating_modes/bridge_mode/country_code_channel_division/country_code_JP/android/test_JP_80Mhz.py
@@ -523,6 +523,16 @@ for sec_modes in setup_params_general5['ssid_modes'].keys():
         setup_params_general5['ssid_modes'][sec_modes][i]['ssid_name'] = setup_params_general5['ssid_modes'][sec_modes][i]['ssid_name'] + "_"+ rand_string
 
 
+@pytest.mark.parametrize(
+    'setup_profiles',
+    [setup_params_general4],
+    indirect=True,
+    scope="class"
+)
+@pytest.mark.usefixtures("setup_profiles")
+class TestBridgeModeConnectSuiteFive(object):
+
+
     @allure.testcase(url="https://telecominfraproject.atlassian.net/browse/WIFI-7290", name="WIFI-7290")
     @pytest.mark.twog
     @pytest.mark.wpa2_personal
@@ -588,6 +598,16 @@ for sec_modes in setup_params_general6['ssid_modes'].keys():
         rand_string = (''.join(random.choices(string.ascii_uppercase +
                                      string.digits, k=N)))+str(int(time.time_ns())%10000)
         setup_params_general6['ssid_modes'][sec_modes][i]['ssid_name'] = setup_params_general6['ssid_modes'][sec_modes][i]['ssid_name'] + "_"+ rand_string
+
+
+@pytest.mark.parametrize(
+    'setup_profiles',
+    [setup_params_general4],
+    indirect=True,
+    scope="class"
+)
+@pytest.mark.usefixtures("setup_profiles")
+class TestBridgeModeConnectSuiteSix(object):
 
 
     @allure.testcase(url="https://telecominfraproject.atlassian.net/browse/WIFI-7290", name="WIFI-7290")
@@ -656,6 +676,16 @@ for sec_modes in setup_params_general7['ssid_modes'].keys():
         setup_params_general7['ssid_modes'][sec_modes][i]['ssid_name'] = setup_params_general7['ssid_modes'][sec_modes][i]['ssid_name'] + "_"+ rand_string
 
 
+@pytest.mark.parametrize(
+    'setup_profiles',
+    [setup_params_general4],
+    indirect=True,
+    scope="class"
+)
+@pytest.mark.usefixtures("setup_profiles")
+class TestBridgeModeConnectSuiteSeven(object):
+
+
     @allure.testcase(url="https://telecominfraproject.atlassian.net/browse/WIFI-7290", name="WIFI-7290")
     @pytest.mark.twog
     @pytest.mark.wpa2_personal
@@ -722,6 +752,16 @@ for sec_modes in setup_params_general8['ssid_modes'].keys():
         setup_params_general8['ssid_modes'][sec_modes][i]['ssid_name'] = setup_params_general8['ssid_modes'][sec_modes][i]['ssid_name'] + "_"+ rand_string
 
 
+@pytest.mark.parametrize(
+    'setup_profiles',
+    [setup_params_general4],
+    indirect=True,
+    scope="class"
+)
+@pytest.mark.usefixtures("setup_profiles")
+class TestBridgeModeConnectSuiteEight(object):
+
+
     @allure.testcase(url="https://telecominfraproject.atlassian.net/browse/WIFI-7290", name="WIFI-7290")
     @pytest.mark.twog
     @pytest.mark.wpa2_personal
@@ -786,6 +826,16 @@ for sec_modes in setup_params_general9['ssid_modes'].keys():
         rand_string = (''.join(random.choices(string.ascii_uppercase +
                                      string.digits, k=N)))+str(int(time.time_ns())%10000)
         setup_params_general9['ssid_modes'][sec_modes][i]['ssid_name'] = setup_params_general9['ssid_modes'][sec_modes][i]['ssid_name'] + "_"+ rand_string
+
+
+@pytest.mark.parametrize(
+    'setup_profiles',
+    [setup_params_general4],
+    indirect=True,
+    scope="class"
+)
+@pytest.mark.usefixtures("setup_profiles")
+class TestBridgeModeConnectSuiteNine(object):
 
 
     @allure.testcase(url="https://telecominfraproject.atlassian.net/browse/WIFI-7290", name="WIFI-7290")
@@ -855,6 +905,16 @@ for sec_modes in setup_params_general10['ssid_modes'].keys():
         setup_params_general10['ssid_modes'][sec_modes][i]['ssid_name'] = setup_params_general10['ssid_modes'][sec_modes][i]['ssid_name'] + "_"+ rand_string
 
 
+@pytest.mark.parametrize(
+    'setup_profiles',
+    [setup_params_general4],
+    indirect=True,
+    scope="class"
+)
+@pytest.mark.usefixtures("setup_profiles")
+class TestBridgeModeConnectSuiteTen(object):
+
+
     @allure.testcase(url="https://telecominfraproject.atlassian.net/browse/WIFI-7290", name="WIFI-7290")
     @pytest.mark.twog
     @pytest.mark.wpa2_personal
@@ -921,10 +981,20 @@ for sec_modes in setup_params_general11['ssid_modes'].keys():
         setup_params_general11['ssid_modes'][sec_modes][i]['ssid_name'] = setup_params_general11['ssid_modes'][sec_modes][i]['ssid_name'] + "_"+ rand_string
 
 
+@pytest.mark.parametrize(
+    'setup_profiles',
+    [setup_params_general8],
+    indirect=True,
+    scope="class"
+)
+@pytest.mark.usefixtures("setup_profiles")
+class TestBridgeModeConnectSuiteEleven(object):
+
+
     @allure.testcase(url="https://telecominfraproject.atlassian.net/browse/WIFI-7290", name="WIFI-7290")
     @pytest.mark.twog
     @pytest.mark.wpa2_personal
-    @pytest.mark.eightyMhz
+    @pytest.mark.plott
     def test_ClientConnect_bridge_wpa2_chn11_80Mhz_JP_2g(self, request, get_vif_state, get_ap_logs,
                                                          get_ToggleAirplaneMode_data, setup_perfectoMobile_android):
         profile_data = setup_params_general11["ssid_modes"]["wpa2_personal"][0]
